APD-123 update rename directoy
