+++ .gitignore
@@ -4,3 +4,4 @@ bnivenv
 __pycache__
 .DS_Store
 dist
+.bnipython.egg-info
