+++ bnipython/lib/api/oneGatePayment.py
@@ -1,6 +1,6 @@
-from bnipython.net.httpClient import HttpClient
-from bnipython.util.utils import generateClientId, generateSignature
-from bnipython.util.response import responseOGP
+from bnipython.lib.net.httpClient import HttpClient
+from bnipython.lib.util.utils import generateClientId, generateSignature
+from bnipython.lib.util.response import responseOGP
 
 
 class OneGatePayment():
